1. Moved PriceProssorSelection in Flight class to be instantiated during fare calculation. 2.Created IFlight service for PriceProcessor to acess required data. 3.Removed PricingService Class as not required. 4.Initialing PricingRulesRepositories at Flight service and injecting it during getTotalFare method call. 5. Modified the test cases accordingly.

+++ src/main/java/airline/Model/PricingRuleModels/FirstClassPricingRuleModel.java
@@ -5,7 +5,7 @@ package airline.Model.PricingRuleModels;
  */
 public class FirstClassPricingRuleModel {
 
-    public float getIncrementalPercentInFare(int dayDiffWithDateOfJourney) {
+    public float getIncrementalPercentInFare(long dayDiffWithDateOfJourney) {
         if((10 - dayDiffWithDateOfJourney) >=0)
             return (10 - dayDiffWithDateOfJourney) * 10f;
